Applied suggestions from code review
